Changed Count Types to int - This has to do with compatibility reasons with unity platforms

+++ Graphene.Library/IEntity.cs
@@ -4,7 +4,7 @@ namespace Graphene
     {
         IGraph Graph { get; }
 
-        ulong Id { get; }
+        int Id { get; }
 
         string Label { get; set; }
